requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2

opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0

python-dotenv==1.0.1
